projectmanagement/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Deprecated==1.2.18
Django==5.2
django-prometheus==2.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
importlib-metadata==6.11.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
opentelemetry-api==1.21.0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-dbapi==0.42b0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-instrumentation-mysql==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-wsgi==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
pillow==11.2.1
prometheus_client==0.21.1
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
typing_extensions==4.13.2
urllib3==2.4.0
wrapt==1.17.2
zipp==3.21.0
